Move trans_const to mir::constant
